Changed int to float

+++ Tutorial/Week 3/Exercise2.py
@@ -1,11 +1,11 @@
 # Develop a program to calculate the average marks of any 5 subjects by documenting your function (using doc string)
 
 # Taking marks from the user
-subject1 = int(input("Enter your marks in subject 1: "))
-subject2 = int(input("Enter your marks in subject 2: "))
-subject3 = int(input("Enter your marks in subject 3: "))
-subject4 = int(input("Enter your marks in subject 4: "))
-subject5 = int(input("Enter your marks in subject 5: "))
+subject1 = float(input("Enter your marks in subject 1: "))
+subject2 = float(input("Enter your marks in subject 2: "))
+subject3 = float(input("Enter your marks in subject 3: "))
+subject4 = float(input("Enter your marks in subject 4: "))
+subject5 = float(input("Enter your marks in subject 5: "))
 
 
 def avg(a, b, c, d, e):
